Move development site workflow to separate component
