+++ app/controllers/projects/edit/geometry-edit.js
@@ -1,5 +1,9 @@
 import Controller from '@ember/controller';
 
 export default class GeometryEditController extends Controller {
-  queryParams = ['mode'];
+  queryParams = ['mode', 'type'];
+
+  type = 'development-site';
+
+  mode = 'lots';
 }
