+++ app/templates/projects/edit/geometry-edit.hbs
@@ -1,3 +1,6 @@
-{{project-geometry-edit model=model mode=mode}}
+{{project-geometry-edit
+  model=model
+  type=type
+  mode=mode}}
 
 {{outlet}}
